Moved automasking and dezingering to IntegrationUtils.py
