+++ src/PyHyperScattering/WPIntegrator.py
@@ -2,6 +2,7 @@ MACHINE_HAS_CUDA = True
 from PyHyperScattering.FileLoader import FileLoader
 from scipy.ndimage import label
 from skimage import filters
+from .IntegrationUtils import automask, remove_zingers
 import os
 import xarray as xr
 import pandas as pd
@@ -25,13 +26,12 @@ try:
 except ImportError:
     warnings.warn('Failed to import Dask, if Dask reduction is desired install pyhyperscattering[performance]',stacklevel=2)
 
-
 class WPIntegrator():
     '''
     Integrator for qx/qy format xarrays using skimage.transform.warp_polar or a custom cuda-accelerated version, warp_polar_gpu
     '''
     
-    def __init__(self,return_cupy=False,force_np_backend=False,use_chunked_processing=False,dezing=True):
+    def __init__(self,return_cupy=False,force_np_backend=False,use_chunked_processing=False,automask=False,dezing=False):
         '''
         Args:
             return_cupy (bool, default False): return arrays as cupy rather than numpy, for further GPU processing
@@ -44,60 +44,9 @@ class WPIntegrator():
             
         self.return_cupy = return_cupy
         self.use_chunked_processing=use_chunked_processing
+	  self.automask = automask
         self.dezing = dezing
-        
-    def mask_image(self, image, max_size = 50):
-        # Create binary mask of the image by thresholding at 0
-        mask = (image <= 0.25)
-
-        # Label the connected regions of the mask
-        labels, num_features = label(mask)
-
-        # Mask out regions that are smaller than max_size
-        for i in range(1, num_features+1):
-            size = np.sum(labels == i)
-            if size <= max_size:
-                mask[labels == i] = False
-
-        # Convert the masked values to NaN
-        image = image.astype(float)
-        image[mask] = np.nan
-
-        return image
-
-    def remove_zingers(self, data_array, threshold1 = 10, threshold2 = 10):        
-        # Compute the mean intensity value across the chi axis for each q
-        mean_intensity = np.nanmean(data_array, axis=1)
-
-        # Compute the standard deviation of the intensity values at each q
-        std_intensity = np.nanstd(data_array, axis=1)
-
-        # Compute the z-score for each intensity value at each q
-        z_score = (data_array - mean_intensity[:, np.newaxis]) / std_intensity[:, np.newaxis]
-
-        # Identify outliers by thresholding the z-score array
-        outliers = np.abs(z_score) > threshold1
-
-        # Iterate over each q coordinate with outliers
-        for i, q in enumerate(outliers):
-            if np.any(q):
-                # Compute the mean intensity value across all chi for this q
-                mean_intensity_q = np.nanmean(data_array[i][~q])
-
-                # Compute the standard deviation of the intensity values at this q
-                std_intensity_q = np.nanstd(data_array[i][~q])
-
-                # Compute the z-score for each intensity value at this q
-                z_score_q = (data_array[i] - mean_intensity_q) / std_intensity_q
-
-                # Identify outliers by thresholding the z-score array for this q
-                outliers_q = np.abs(z_score_q) > threshold2
-
-                # Mask the outliers with NaN values
-                data_array[i][outliers_q] = np.nan
-
-        return data_array
-    
+            
     def warp_polar_gpu(self,image, center=None, radius=None, output_shape=None, **kwargs):
         """
         Function to emulate warp_polar in skimage.transform on the GPU. Not all
@@ -171,9 +120,15 @@ class WPIntegrator():
         else:
             TwoD = skimage.transform.warp_polar(img_to_integ,center=(center_x,center_y), radius = np.sqrt((img_to_integ.shape[0] - center_x)**2 + (img_to_integ.shape[1] - center_y)**2))
 
-        TwoD_dezinged = self.remove_zingers(np.array(TwoD))
+	  if self.dezing:
+        	TwoD_dezinged = self.remove_zingers(np.array(TwoD))
+	  else:
+		TwoD_dezinged = np.array(TwoD)
         
-        TwoD_dezinged_masked = self.mask_image(TwoD_dezinged)
+	  if self.automask:
+        	TwoD_dezinged_masked = self.automask(TwoD_dezinged)
+	  else:
+		TwoD_dezinged_masked = TwoD_dezinged
         
         xarr = TwoD_dezinged_masked
